refactor: Move types to own directory
