--- src/background/types.ts
@@ -1,24 +1,0 @@
-interface OperaWorkspaceEntry {
-    workspaceId: string;
-    workspaceName: string;
-}
-
-export interface OperaTab extends chrome.tabs.Tab, OperaWorkspaceEntry {}
-
-export interface BookmarksBar {
-    id: string;
-    title: string;
-}
-
-export interface BookmarksBarOpera extends BookmarksBar {
-    workspaceId: string;
-}
-
-export interface BookmarksBarPopup extends BookmarksBar {
-    isActive: boolean;
-    isEdited: boolean;
-}
-
-export interface BookmarksBarRemoveCandidate extends BookmarksBarPopup {
-    index: number;
-}
